KnowledgeHarvester/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pypdf==4.0.1
markdownify==0.11.6
tqdm==4.66.2
python-dotenv==1.0.1
pyyaml==6.0.1

